+++ backend/store/etcd/initialization.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"path"
 
-	"github.com/coreos/etcd/clientv3"
-	"github.com/coreos/etcd/clientv3/concurrency"
+	"go.etcd.io/etcd/clientv3"
+	"go.etcd.io/etcd/clientv3/concurrency"
 	"github.com/sensu/sensu-go/backend/store"
 )