+++ backend/store/v2/etcdstore/store.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/coreos/etcd/clientv3"
+	"go.etcd.io/etcd/clientv3"
 	"github.com/gogo/protobuf/proto"
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	corev3 "github.com/sensu/sensu-go/api/core/v3"
@@ -257,7 +257,7 @@ func (s *Store) Get(req storev2.ResourceRequest) (*storev2.Wrapper, error) {
 	ctx := req.Context
 	var resp *clientv3.GetResponse
 	err := Backoff(ctx).Retry(func(n int) (done bool, err error) {
-		resp, err = s.client.Get(ctx, key, clientv3.WithLimit(1))
+		resp, err = s.client.Get(ctx, key, clientv3.WithLimit(1), clientv3.WithSerializable())
 		return RetryRequest(n, err)
 	})
 	if err != nil {
@@ -349,4 +349,21 @@ func (s *Store) List(req storev2.ResourceRequest, pred *store.SelectionPredicate
 		pred.Continue = ""
 	}
 	return result, nil
+}
+
+func (s *Store) Exists(req storev2.ResourceRequest) (bool, error) {
+	key := StoreKey(req)
+	if err := req.Validate(); err != nil {
+		return false, &store.ErrNotValid{Err: err}
+	}
+	ctx := req.Context
+	var resp *clientv3.GetResponse
+	err := Backoff(ctx).Retry(func(n int) (done bool, err error) {
+		resp, err = s.client.Get(ctx, key, clientv3.WithLimit(1), clientv3.WithSerializable(), clientv3.WithCountOnly())
+		return RetryRequest(n, err)
+	})
+	if err != nil {
+		return false, err
+	}
+	return resp.Count > 0, nil
 }