+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Added
+- Added Prometheus transformer for extracting metrics from check output
+using the Prometheus Exposition Text Format.
+
 ## [6.0.0] - 2020-07-23
 
 ### Added
@@ -17,8 +21,6 @@ enabled.
 ### Fixed
 - Clarifies wording around a secret provider error message.
 - Logs and returns an error if a mutator cannot be found.
-- Errors produced in the agent by assets, check validation, token substitution,
-and event unmarshaling are logged once again.
 - User-Agent header is only set on new client creation rather than on each
 request.
 - Fixed a bug where highly concurrent event filtering could result in a panic.
@@ -55,6 +57,17 @@ characters.
 - Enriches output of `sensuctl asset add` with help usage for how to use the runtime asset.
 - Unless the entity is a proxy entity, updates to entities now ignore state
   related fields.
+- Upgraded embedded etcd to 3.4.10
+
+## [5.21.1] - 2020-08-05
+
+### Changed
+- Improved agent websocket connection logging.
+
+### Fixed
+- Fixed potential Web UI crash when fetching events in namespace with > 1000 events.
+- Errors produced in the agent by assets, check validation, token substitution,
+and event unmarshaling are logged once again.
 
 ## [5.21.0] - 2020-06-10